Add files via upload
Project-Team-Portal
Android App

Aim -: The aim of this project is to create Group for Project, Where the students as well as teachers can post their own Project of events for which they require team members to work on the projects and we are creating a portal for group formation for the projects where we can post our projects ideas and after the posting the project ideas the interested students participate the project group and start discussing the project. Our aim is to provide a platform where you can find good projects with better team and knowledge about the project. This Android Project helps the students to find their team members for the project group, the team portal helps to create their own projects and discuss with other team members online and they can post their own ideas and build a good Project Team.

Database used: Firebase Realtime Database Firebase functions are also used for sending peer to peer notifications